.claude/requirements.txt
# Audio/Video to Text Converter dependencies
vertexai
python-dotenv
ffmpeg-python

# Google Calendar Today dependencies
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client